requirements.txt
fastapi
uvicorn[standard]
yt-dlp
openai-whisper
python-multipart
torch
snscrape
tweepy
python-dotenv
requests
beautifulsoup4
pydantic